requirements.txt
opencv-python
pillow
deepface
google-generativeai
pyttsx3
tf-keras
